Create a longDiv package and added a GenerateHtml method.
LongDiv was initally developed as a binary to develop the algorithm.
LongDiv has now been moved to its own package and the binary code moved
to division.go which imports the longDiv package. A test suite was added
for the longDiv package.
The PrintWorking method in webserver for LongMult was name changed to
GenerateHtml for consistency with the longDiv package.