Added remaining files from GKE Cloud Console working space.
These include the bash script used to build and deploy the helloweb app.
The yaml manifests which define the deployment except the ingress yaml (which contains secrets).
The Dockerfile which creates the container.
The bash scripts used to launch the helloweb app and to clean up afterwards.
The utility script used to download files if they are edited locally.